chore: move to a folder structure
